+++ app/pages/participant/tile.tsx
@@ -1,0 +1,114 @@
+import { isTrackReferencePlaceholder } from '@/app/devices/video_container';
+import {
+  AudioTrack,
+  ConnectionQualityIndicator,
+  isTrackReference,
+  LockLockedIcon,
+  ParticipantName,
+  ParticipantPlaceholder,
+  ParticipantTile,
+  ParticipantTileProps,
+  PinState,
+  ScreenShareIcon,
+  TrackMutedIndicator,
+  TrackReferenceOrPlaceholder,
+  useEnsureTrackRef,
+  useFeatureContext,
+  useIsEncrypted,
+  useMaybeLayoutContext,
+  VideoTrack,
+} from '@livekit/components-react';
+import { Track } from 'livekit-client';
+import React from 'react';
+
+export function ParticipantItem({ trackRef }: ParticipantTileProps) {
+  const trackReference = useEnsureTrackRef(trackRef);
+  const isEncrypted = useIsEncrypted(trackReference.participant);
+  const layoutContext = useMaybeLayoutContext();
+  const autoManageSubscription = useFeatureContext()?.autoSubscription;
+  const handleSubscribe = React.useCallback(
+    (subscribed: boolean) => {
+      if (
+        trackReference.source &&
+        !subscribed &&
+        layoutContext &&
+        layoutContext.pin.dispatch &&
+        isTrackReferencePinned(trackReference, layoutContext.pin.state)
+      ) {
+        layoutContext.pin.dispatch({ msg: 'clear_pin' });
+      }
+    },
+    [trackReference, layoutContext],
+  );
+
+  return (
+    <ParticipantTile>
+      {isTrackReference(trackReference) &&
+      (trackReference.publication?.kind === 'video' ||
+        trackReference.source === Track.Source.Camera ||
+        trackReference.source === Track.Source.ScreenShare) ? (
+        <VideoTrack
+          trackRef={trackReference}
+          onSubscriptionStatusChanged={handleSubscribe}
+          manageSubscription={autoManageSubscription}
+        />
+      ) : (
+        isTrackReference(trackReference) && (
+          <AudioTrack trackRef={trackReference} onSubscriptionStatusChanged={handleSubscribe} />
+        )
+      )}
+      <div className="lk-participant-placeholder">
+        <ParticipantPlaceholder />
+      </div>
+      <div className="lk-participant-metadata">
+        <div className="lk-participant-metadata-item">
+          {trackReference.source === Track.Source.Camera ? (
+            <>
+              {isEncrypted && <LockLockedIcon style={{ marginRight: '0.25rem' }} />}
+              <TrackMutedIndicator
+                trackRef={{
+                  participant: trackReference.participant,
+                  source: Track.Source.Microphone,
+                }}
+                show={'muted'}
+              ></TrackMutedIndicator>
+              <ParticipantName />
+            </>
+          ) : (
+            <>
+              <ScreenShareIcon style={{ marginRight: '0.25rem' }} />
+              <ParticipantName>&apos;s screen</ParticipantName>
+            </>
+          )}
+        </div>
+        <ConnectionQualityIndicator className="lk-participant-metadata-item" />
+      </div>
+    </ParticipantTile>
+  );
+}
+
+export function isTrackReferencePinned(
+  trackReference: TrackReferenceOrPlaceholder,
+  pinState: PinState | undefined,
+): boolean {
+  if (typeof pinState === 'undefined') {
+    return false;
+  }
+  if (isTrackReference(trackReference)) {
+    return pinState.some(
+      (pinnedTrackReference) =>
+        pinnedTrackReference.participant.identity === trackReference.participant.identity &&
+        isTrackReference(pinnedTrackReference) &&
+        pinnedTrackReference.publication.trackSid === trackReference.publication.trackSid,
+    );
+  } else if (isTrackReferencePlaceholder(trackReference)) {
+    return pinState.some(
+      (pinnedTrackReference) =>
+        pinnedTrackReference.participant.identity === trackReference.participant.identity &&
+        isTrackReferencePlaceholder(pinnedTrackReference) &&
+        pinnedTrackReference.source === trackReference.source,
+    );
+  } else {
+    return false;
+  }
+}