+++ lib/std/index.ts
@@ -38,3 +38,11 @@ export function src(url: string): string {
   }
   return `${prefix}${url}`;
 }
+
+export function connect_endpoint() {
+  let prefix = process.env.NEXT_PUBLIC_BASE_PATH;
+  if (!prefix || prefix === '' || prefix === '/') {
+    return process.env.NEXT_PUBLIC_CONN_DETAILS_ENDPOINT ?? '/api/connection-details';
+  }
+  return `${prefix}${process.env.NEXT_PUBLIC_CONN_DETAILS_ENDPOINT ?? '/api/connection-details'}`;
+}